Músicas em processamento e filtrar por categoria
Músicas em processamento e filtrar por categoria

+++ API/api/components/listaReproducao/listaReproducaoController.js
@@ -8,7 +8,7 @@ const ytdl = require('ytdl-core');
 
 
 exports.createList = async (req, res) => {
-    console.log(req.body)
+
     let serverResponse = { status: "Lista não criada", response: {} }
     //variável que guarda a query à base de dados
     var lista;
@@ -17,25 +17,38 @@ exports.createList = async (req, res) => {
     var existsNomeListaConfirm;
     //variável que recolhe o parâmetro enviado na request
     var dadosLista = {
-        nomeLista:req.body.nomeLista,
-        userFK:req.body.userFK
-        };
-    await listaReproducaoService.getList(dadosLista.userFK,dadosLista.nomeLista).then(url =>existsNomeLista=url).catch(err => console.log(err));
-    if(existsNomeLista ==null){
-        await listaReproducaoService.createList(dadosLista).then(url => lista = url).catch(err => console.log(err));
+        nomeLista: req.body.nomeLista,
+        userFK: req.body.userFK
+    };
+    var token = req.headers['x-access-token'];
+    if (!token) {
+        serverResponse = { status: "Nao está autenticado", response: {} }
+        return res.send(serverResponse);
+    }
+    try {
+        jwt.verify(token, 'secret');
 
-        if (lista != null) {
-            await listaReproducaoService.getList(dadosLista.userFK,dadosLista.nomeLista).then(url =>existsNomeListaConfirm=url).catch(err => console.log(err));
+        await listaReproducaoService.getList(dadosLista.userFK, dadosLista.nomeLista).then(url => existsNomeLista = url).catch(err => console.log(err));
+        if (existsNomeLista == null) {
+            await listaReproducaoService.createList(dadosLista).then(url => lista = url).catch(err => console.log(err));
 
-            serverResponse = { status: "Lista criada", response: existsNomeListaConfirm }
+            if (lista != null) {
+                await listaReproducaoService.getList(dadosLista.userFK, dadosLista.nomeLista).then(url => existsNomeListaConfirm = url).catch(err => console.log(err));
+
+                serverResponse = { status: "Lista criada", response: existsNomeListaConfirm }
+            }
         }
-    }
-    else{
-        serverResponse = { status: "Já existe uma lista com o nome escolhido", response: {} }
+        else {
+            serverResponse = { status: "Já existe uma lista com o nome escolhido", response: {} }
 
+        }
+
+        return res.send(serverResponse);
     }
-    
-    return res.send(serverResponse);
+    catch (err) {
+        serverResponse = { status: "Nao está autenticado | token expirou", response: {} }
+        return res.send(serverResponse);
+    }
 }
 
 exports.deleteList = async (req, res) => {
@@ -43,13 +56,25 @@ exports.deleteList = async (req, res) => {
     //variável que guarda a query à base de dados
     var lista;
     var listaID = req.params.id;
-    //variável que recolhe o parâmetro enviado na request
 
-    await listaReproducaoService.deleteList(listaID).then(url => lista = url).catch(err => console.log(err));
-    if (lista != null) {
-        serverResponse = { status: "Lista Eliminada com sucesso", response: lista }
+    var token = req.headers['x-access-token'];
+    if (!token) {
+        serverResponse = { status: "Nao está autenticado", response: {} }
+        return res.send(serverResponse);
     }
-    return res.send(serverResponse);
+    try {
+        jwt.verify(token, 'secret');
+
+        await listaReproducaoService.deleteList(listaID).then(url => lista = url).catch(err => console.log(err));
+        if (lista != null) {
+            serverResponse = { status: "Lista Eliminada com sucesso", response: lista }
+        }
+        return res.send(serverResponse);
+    }
+    catch (err) {
+        serverResponse = { status: "Nao está autenticado | token expirou", response: {} }
+        return res.send(serverResponse);
+    }
 }
 
 exports.getListUser = async (req, res) => {
@@ -57,11 +82,22 @@ exports.getListUser = async (req, res) => {
     //variável que guarda a query à base de dados
     var listas;
     var userFK = req.body.userFK;
-    //variável que recolhe o parâmetro enviado na request
+    var token = req.headers['x-access-token'];
+
+    if (!token) {
+        serverResponse = { status: "Nao está autenticado", response: {} }
+        return res.send(serverResponse);
+    }
+    try{
+        jwt.verify(token, 'secret');
 
     await listaReproducaoService.getListUser(userFK).then(url => listas = url).catch(err => console.log(err));
     if (listas != null) {
         serverResponse = { status: "Lista Listadas com sucesso", response: listas }
     }
     return res.send(serverResponse);
+}catch(err){
+    serverResponse = { status: "Nao está autenticado | token expirou", response: {} }
+    return res.send(serverResponse);
+}
 }
